Move `Client/ServerCacheControl` to `common`
Motivation:

Both `ClientCacheControl` and `ServerCacheControl` can be used in either
client or server side and thus they should be in the `common` package.

Modifications:

- Move `Client/ServerCacheControl` to `common`.

Result:

- Fixes #1813
- Better organization
