+++ core/src/main/java/com/linecorp/armeria/common/ServerCacheControlBuilder.java
@@ -13,14 +13,11 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package com.linecorp.armeria.server;
+package com.linecorp.armeria.common;
 
 import java.time.Duration;
 
 import javax.annotation.Nullable;
-
-import com.linecorp.armeria.client.ClientCacheControlBuilder;
-import com.linecorp.armeria.common.CacheControlBuilder;
 
 /**
  * Creates a new {@link ServerCacheControl} using the builder pattern.
