+++ core/src/test/java/com/linecorp/armeria/common/HttpHeadersTest.java
@@ -36,8 +36,6 @@ import java.util.TimeZone;
 
 import org.junit.Test;
 
-import com.linecorp.armeria.server.ServerCacheControl;
-
 public class HttpHeadersTest {
 
     @Test
