+++ core/src/test/java/com/linecorp/armeria/server/file/HttpFileServiceAdditionalHeadersTest.java
@@ -24,8 +24,8 @@ import com.linecorp.armeria.client.HttpClient;
 import com.linecorp.armeria.common.AggregatedHttpResponse;
 import com.linecorp.armeria.common.HttpHeaderNames;
 import com.linecorp.armeria.common.HttpStatus;
+import com.linecorp.armeria.common.ServerCacheControl;
 import com.linecorp.armeria.server.ServerBuilder;
-import com.linecorp.armeria.server.ServerCacheControl;
 import com.linecorp.armeria.testing.junit4.server.ServerRule;
 
 public class HttpFileServiceAdditionalHeadersTest {
